advertisingForum/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.13
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
pkg_resources==0.0.0
psycopg2==2.7.7
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
whitenoise==5.3.0
